Fix GtfsToAndroidSqlite
