--- GtfsToAndroidSqlite/.idea/description.html
@@ -1,1 +1,0 @@
-<html>Simple <b>Java</b> application that includes a class with <code>main()</code> method</html>
